src/python/pricing_intelligence/service.py
#!/usr/bin/env python3
"""
Pricing Intelligence Service
Web scraping and competitive pricing analysis
"""

import asyncio
import logging
from typing import Dict, List, Any
from datetime import datetime
import requests
from bs4 import BeautifulSoup
import psycopg2
from psycopg2.extras import RealDictCursor

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class PricingIntelligenceService:
    """Pricing intelligence and competitive analysis"""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.pg_conn = None
        self._initialize_connections()
        logger.info("Pricing Intelligence service initialized")
    
    def _initialize_connections(self):
        """Initialize database connection"""
        try:
            self.pg_conn = psycopg2.connect(
                host=self.config.get('postgres_host', 'localhost'),
                port=self.config.get('postgres_port', 5432),
                database=self.config.get('postgres_db', 'ai_engine'),
                user=self.config.get('postgres_user', 'weedgo'),
                password=self.config.get('postgres_password', 'your_password_here')
            )
            logger.info("Connected to PostgreSQL for pricing intelligence")
        except Exception as e:
            logger.error(f"Failed to connect to database: {e}")
    
    async def analyze_pricing(self, product_ids: List[str]) -> Dict[str, Any]:
        """Analyze pricing for given products"""
        try:
            analysis_results = []
            
            for product_id in product_ids:
                # Get current product pricing
                current_price = await self._get_current_price(product_id)
                
                # Get competitor pricing
                competitor_prices = await self._get_competitor_prices(product_id)
                
                # Calculate market statistics
                market_stats = self._calculate_market_stats(competitor_prices)
                
                # Generate pricing recommendation
                recommendation = self._generate_pricing_recommendation(
                    current_price, market_stats
                )
                
                analysis_results.append({
                    'product_id': product_id,
                    'current_price': current_price,
                    'market_average': market_stats['average'],
                    'market_min': market_stats['min'],
                    'market_max': market_stats['max'],
                    'recommended_price': recommendation['price'],
                    'competitive_position': recommendation['position'],
                    'optimization_score': recommendation['score'],
                    'insights': recommendation['insights']
                })
            
            return {
                'analysis_results': analysis_results,
                'analysis_timestamp': datetime.now().isoformat(),
                'market_summary': self._generate_market_summary(analysis_results)
            }
            
        except Exception as e:
            logger.error(f"Pricing analysis error: {e}")
            return {'error': str(e)}
    
    async def scrape_competitor_prices(self, competitors: List[str]) -> Dict[str, Any]:
        """Scrape competitor pricing data"""
        try:
            scraping_results = []
            
            for competitor in competitors:
                # Simulate web scraping (in real implementation, this would scrape actual sites)
                result = await self._scrape_competitor_site(competitor)
                scraping_results.append(result)
            
            # Store results in database
            await self._store_competitor_data(scraping_results)
            
            return {
                'scraping_results': scraping_results,
                'total_products_found': sum(r['products_found'] for r in scraping_results),
                'scraping_timestamp': datetime.now().isoformat(),
                'next_scrape_scheduled': datetime.now().isoformat()  # Would be scheduled
            }
            
        except Exception as e:
            logger.error(f"Competitor scraping error: {e}")
            return {'error': str(e)}
    
    async def _get_current_price(self, product_id: str) -> float:
        """Get current price for a product"""
        try:
            if not self.pg_conn:
                return 25.99  # Default price
            
            cursor = self.pg_conn.cursor()
            cursor.execute("""
                SELECT price FROM cannabis_data.product_pricing 
                WHERE product_id = %s 
                ORDER BY created_at DESC 
                LIMIT 1
            """, (product_id,))
            
            result = cursor.fetchone()
            cursor.close()
            
            return result[0] if result else 25.99
            
        except Exception as e:
            logger.error(f"Error getting current price: {e}")
            return 25.99
    
    async def _get_competitor_prices(self, product_id: str) -> List[float]:
        """Get competitor prices for a product"""
        try:
            if not self.pg_conn:
                # Simulate competitor prices
                return [22.99, 24.50, 26.99, 28.00, 23.75]
            
            cursor = self.pg_conn.cursor()
            cursor.execute("""
                SELECT price FROM ai_engine.competitor_pricing 
                WHERE product_match_id = %s 
                AND scraped_at > NOW() - INTERVAL '7 days'
            """, (product_id,))
            
            results = cursor.fetchall()
            cursor.close()
            
            prices = [result[0] for result in results] if results else [22.99, 24.50, 26.99, 28.00, 23.75]
            return prices
            
        except Exception as e:
            logger.error(f"Error getting competitor prices: {e}")
            return [22.99, 24.50, 26.99, 28.00, 23.75]
    
    def _calculate_market_stats(self, prices: List[float]) -> Dict[str, float]:
        """Calculate market statistics"""
        if not prices:
            return {'average': 25.0, 'min': 20.0, 'max': 30.0}
        
        return {
            'average': sum(prices) / len(prices),
            'min': min(prices),
            'max': max(prices),
            'median': sorted(prices)[len(prices) // 2]
        }
    
    def _generate_pricing_recommendation(self, current_price: float, 
                                       market_stats: Dict[str, float]) -> Dict[str, Any]:
        """Generate pricing recommendation"""
        market_avg = market_stats['average']
        
        # Determine competitive position
        if current_price < market_avg * 0.9:
            position = 'below_market'
            recommended_price = market_avg * 0.95
            insights = ['Consider increasing price to capture more value']
        elif current_price > market_avg * 1.1:
            position = 'above_market'
            recommended_price = market_avg * 1.05
            insights = ['Consider lowering price to increase competitiveness']
        else:
            position = 'competitive'
            recommended_price = current_price
            insights = ['Price is well positioned in the market']
        
        # Calculate optimization score
        price_diff = abs(current_price - market_avg) / market_avg
        optimization_score = max(0, 1 - price_diff)
        
        return {
            'price': round(recommended_price, 2),
            'position': position,
            'score': round(optimization_score, 3),
            'insights': insights
        }
    
    def _generate_market_summary(self, analysis_results: List[Dict]) -> Dict[str, Any]:
        """Generate overall market summary"""
        total_products = len(analysis_results)
        
        positions = [r['competitive_position'] for r in analysis_results]
        position_counts = {
            'below_market': positions.count('below_market'),
            'competitive': positions.count('competitive'),
            'above_market': positions.count('above_market')
        }
        
        avg_optimization_score = sum(r['optimization_score'] for r in analysis_results) / total_products
        
        return {
            'total_products_analyzed': total_products,
            'position_distribution': position_counts,
            'average_optimization_score': round(avg_optimization_score, 3),
            'overall_competitiveness': 'strong' if avg_optimization_score > 0.8 else 'moderate' if avg_optimization_score > 0.6 else 'needs_improvement'
        }
    
    async def _scrape_competitor_site(self, competitor: str) -> Dict[str, Any]:
        """Scrape a competitor website (simulated)"""
        # In a real implementation, this would actually scrape competitor sites
        # For now, we'll simulate the results
        
        await asyncio.sleep(0.1)  # Simulate scraping delay
        
        # Simulate different competitors with different characteristics
        base_products = 100
        if 'premium' in competitor.lower():
            products_found = base_products + 50
            avg_price_diff = '+15.2%'
        elif 'budget' in competitor.lower():
            products_found = base_products + 20
            avg_price_diff = '-8.5%'
        else:
            products_found = base_products + (hash(competitor) % 30)
            avg_price_diff = f"{((hash(competitor) % 20) - 10):+.1f}%"
        
        return {
            'competitor': competitor,
            'products_found': products_found,
            'successful_scrapes': products_found,
            'failed_scrapes': 5,
            'avg_price_difference': avg_price_diff,
            'last_updated': datetime.now().isoformat(),
            'data_quality_score': 0.85 + (hash(competitor) % 15) / 100
        }
    
    async def _store_competitor_data(self, scraping_results: List[Dict]):
        """Store competitor data in database"""
        try:
            if not self.pg_conn:
                return
            
            cursor = self.pg_conn.cursor()
            
            for result in scraping_results:
                cursor.execute("""
                    INSERT INTO ai_engine.competitor_pricing 
                    (competitor_name, competitor_url, product_name, brand, price, 
                     in_stock, scraped_at, data_quality_score)
                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s)
                    ON CONFLICT DO NOTHING
                """, (
                    result['competitor'],
                    f"https://{result['competitor']}.com",
                    'Sample Product',
                    'Sample Brand',
                    25.99,
                    True,
                    datetime.now(),
                    result['data_quality_score']
                ))
            
            self.pg_conn.commit()
            cursor.close()
            
        except Exception as e:
            logger.error(f"Error storing competitor data: {e}")
            if self.pg_conn:
                self.pg_conn.rollback()

def create_pricing_intelligence_service(config: Dict[str, Any]) -> PricingIntelligenceService:
    """Create pricing intelligence service"""
    return PricingIntelligenceService(config)