src/python/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# gRPC and Protocol Buffers
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Web Framework (for health checks and metrics)
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Machine Learning Frameworks
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.2.0

# NLP and Language Support
langdetect==1.0.9
polyglot==16.7.4
# Multi-language tokenizers
jieba==0.42.1  # Chinese
arabic-reshaper==3.0.0  # Arabic
python-bidi==0.4.2  # Arabic RTL

# Computer Vision and OCR
opencv-python==4.9.0.80
Pillow==10.2.0
easyocr==1.7.1
paddlepaddle==2.5.2
paddleocr==2.7.0.3
pytesseract==0.3.10

# Face Recognition
face-recognition==1.3.0
deepface==0.0.79
insightface==0.7.3
onnx==1.15.0
onnxruntime-gpu==1.16.3

# Web Scraping
beautifulsoup4==4.12.2
scrapy==2.11.0
playwright==1.40.0
selenium==4.16.0
undetected-chromedriver==3.5.4
requests-html==0.10.0
lxml==5.0.0

# Vector Database
pymilvus==2.3.4
faiss-gpu==1.7.2
chromadb==0.4.22

# Traditional Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Caching
redis==5.0.1
hiredis==2.3.2

# ML Operations
mlflow==2.9.2
optuna==3.5.0  # Hyperparameter optimization
ray[tune]==2.9.0  # Distributed training
wandb==0.16.2  # Experiment tracking

# Data Processing
pyarrow==14.0.2
dask==2023.12.1
modin[ray]==0.25.0

# Time Series Analysis
prophet==1.1.5
statsmodels==0.14.1

# Monitoring and Logging
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-grpc==0.43b0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.92.2

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3

# Security
cryptography==41.0.7
pycryptodome==3.19.1

# Utilities
tqdm==4.66.1
click==8.1.7
pyyaml==6.0.1
python-json-logger==2.0.7
tenacity==8.2.3  # Retry logic
schedule==1.2.0  # Task scheduling

# Performance
numba==0.58.1  # JIT compilation
cython==3.0.7

# Cannabis-specific NLP
spacy==3.7.2
# Download models with: python -m spacy download en_core_web_lg

# Development
ipython==8.19.0
jupyter==1.0.0
jupyterlab==4.0.10

# Type hints
types-redis==4.6.0.20240106
types-requests==2.31.0.20240106
types-PyYAML==6.0.12.12